refactor(ble): 成功把BLE剔出main
用了一个target_compile_options(${COMPONENT_LIB} PRIVATE -Wno-unused-const-variable)来关掉没使用的错误
