+++ ble_detect_hid/components/bsp_ble/esp_hidd_prf_api.h
