+++ ble_detect_hid/components/bsp_ble/hid_dev.h
