--- ble_detect_hid/main/bsp_ble/CMakeLists.txt
@@ -1,5 +1,0 @@
-idf_component_register(
-    SRCS "ble_module.c"
-    INCLUDE_DIRS "."
-    REQUIRES bsp_button driver
-)
